Add files via upload
Created my own script file.
The Goal?
Running dizqutv under systemctl in the background on Raspberry Pi

On Pi, it'll get stuck trying to remove a file if it existed. Wait 15-20 seconds to cancel to get out of it. Created service file will still be added and enabling the service.